Move before_send specs into existing spec/config/initializers/sentry_spec.rb
Review feedback: initializer specs live under spec/config/initializers/
and a sentry_spec.rb already existed there.
